Move components from dashboard to section dir
