--- src/components/dashboard/Dashboard.scss
@@ -1,12 +1,0 @@
-.Dashboard {
-    position: relative;
-    z-index: 2;
-    background-color: #FFF;
-    box-shadow: 0 2px 2px rgba(0,0,0,0.05);
-    overflow: hidden;
-    &:after {
-        content: "";
-        display: block;
-        clear: both;
-    }
-}
