+++ src/components/pages/dashboard/sections/AssignmentList.scss
